Add submodules and build scripts
Add vaccelrt, firecracker and virtio-accel submodules in the project and
a set of scripts to build all the components. At the moment it is
missing a project and the corresponding script to build a rootfs image
for firecracker.

Also, we need to add logic in the future for building the virtio module
and appropriate kernel for QEMU and also for ARM64.

Signed-off-by: Babis Chalios <mail@bchalios.io>