Build Firecracker rootfs
Add scripts that build a rootfs image that works (at least) for
Firecracker. The script uses a Dockerfile and BUILDKIT to extract a base
image (at the moment we only support `ubuntu:latest`). Then, it adds in
the image everything needed to run a vaccel example inside a Firecracker
VM. That includes: libvaccel, the virtio module preloaded in the guest
kernel, a classification example and a set of images. Finally, all the
environment variables needed by Firecracker are set.

Signed-off-by: Babis Chalios <mail@bchalios.io>