ci/cd: e2e test of vaccel
Set up basic end-to-end testing for Vaccel. The test is run through
Github actions. It uses built components from the submodules and builds
a rootfs image to use with Firecracker. The rootfs includes libvaccel,
the virtio module and the example(s) to be run as an e2e test.

Once the rootfs is built, a Firecracker VM is launched and the tests are
run.

Signed-off-by: Babis Chalios <mail@bchalios.io>