ci/cd: Unify jobs to avoid using artifacts
We have been hitting the Github limits for artifact uploading, which
causes workflow run to fail. To avoid using the artifact
uploading/downloading step we unify the step that downloads objects from
MinIO with the one that runs the test.

Moreover, we use a custom action to download objects from MinIO, so that
we avoid hitting the docker-in-docker issues of our self-hosted runner.

Signed-off-by: Babis Chalios <mail@bchalios.io>