Add vaccel gRPC protocol and vsock plugin
Add a protobuf protocol for vAccel, exposing our API over gRPC. At the
same time, we implement a vsock-based plugin for vaccelrt. This will
serve as an alternative to our custom virtio implementation. The plugin
uses the vaccel gRPC protocol to communicate with the host where a
"vAccel-server" agent listens for vAccel requests, handles them and
responds to the guest application.

Add, as well, an agent implementation for the gRPC API and a CLI client
for testing it.

Signed-off-by: Babis Chalios <mail@bchalios.io>