bindings: keep them outside vaccelrt
It is much easier to handle bindings outside of the vaccelrt repo.
Instead, we add vaccelrt as a submodule inside the rust-bindings repo
(which now is a vaccel submodule).

The vaccel-bindings crate uses the vaccelrt submodule to build itself,
so that it does not require to have a vaccelrt installation already
present. This facilitates building vaccel-bindings themselves and all
the crates that use it.

Signed-off-by: Babis Chalios <mail@bchalios.io>