Introduce API for TensorFlow inference
We introduce an API for running Machine Learning inference using
TensorFlow models. We define the API in vAccel runtime and implement it
in the `noop` plugin. Moreover, we implement the API in the rust
bindings crate.

Also, add a new plugin that implements the TensorFlow API using the
TensorFlow library. The plugin consumes the C TensorFlow API.

Extend the gRPC model of vAccel to describe the requests for the
TensorFlow API and the corresponding implementations for the vsock
plugin and the vAccel agent.

Signed-off-by: Babis Chalios <mail@bchalios.io>