fc: fix firecracker with latest vAccel bindings
Introducing the resources API in vAccel bindings had actually broken
Firecracker. The `vaccel_session` struct now includes a pointer to void
which renders the struct non-`Send`. This means that we have to
implement the Trait for being able to use. We add this (dummy)
implementation in Firecracker, until this is fixed in the bindings
crate.

Moreover, the bindings crate is building vaccelrt itself, so we do not
need to build the runtime inside the container image.

Signed-off-by: Babis Chalios <mail@bchalios.io>