Add a workflow that builds a Firecracker rootfs
Add a workflow that builds a Firecracker root file system every time a
change in the corresponding Dockerfile occurs and uploads it to s3.

This will help us in the future to speed-up e2e testing, since we will
not need to build the rootfs every time. Instead we will just need to
install the vAccel binaries in the rootfs.

Signed-off-by: Babis Chalios <mail@bchalios.io>