ci/cd: do not create a release with pushes to main
Instead for both pushes to main and PRs create an artifact that includes
the base root filesystem with vAccel installed in it and upload it to s3
under the name of the branch.

Also, we change the way we stop Firecracker. Instead of killing it, we
reboot it, which is the correct way of cleanly stopping it. We need to
do that before creating the artifact so the rootfs.img is not used,
while we tar it.

Finally, we stop properly the vAccel agent (by killing it).

Signed-off-by: Babis Chalios <mail@bchalios.io>